🩹 fix 토스트 이미지 에디터 UI 조정
